MAINT: Move functionality from fluid solver scheme to abstract superclass
